+++ src/App.jsx
@@ -1,8 +1,7 @@
 import "./App.css";
-import { About } from "./components/About";
-import { Tech } from "./components/Tech";
-import { Projects } from "./components/Projects";
-
+import { About } from "./components/About/About";
+import { Tech } from "./components/Tech/Tech";
+import { Projects } from "./components/Projects/Projects";
 
 export const App = () => {
   return (
