+++ src/components/Tech/Tech.jsx
